# openapi/protocol-union-api/openapi.yaml
openapi: 3.0.3
info:
  title: OpenAPI definition
  version: v0.1
paths:
...
  /v0.1/ownerships/{ownershipId}:
    get:
      tags:
        - ownership-controller
      operationId: getOwnershipById
      parameters:
        - name: "ownershipId"
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ownership"
        '404':
          $ref: "#/components/responses/EntityNotFound"
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerError"
  /v0.1/ownerships/byItem:
    get:
      tags:
        - ownership-controller
      operationId: getOwnershipsByItem
      parameters:
        - name: "contract"
          in: query
          required: true
          schema:
            type: string
        - name: "tokenId"
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/BigInteger"
        - name: "continuation"
          in: query
          required: false
          schema:
            type: string
        - name: "size"
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ownerships"
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerError"
  /v0.1/ownerships/all:
    get:
      tags:
        - ownership-controller
      operationId: getAllOwnerships
      parameters:
        - name: "blockchains"
          in: query
          required: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/Blockchain"
        - name: "continuation"
          in: query
          required: false
          schema:
            type: string
        - name: "size"
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ownerships"
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerError"

  ######################## Order API ########################
  /v0.1/orders/{id}:
    get:
      tags:
        - order-controller
      operationId: getOrderById
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        '404':
          $ref: "#/components/responses/EntityNotFound"
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerError"
  /v0.1/orders/{id}/updateMakeStock:
    get:
      tags:
        - order-controller
      operationId: "updateOrderMakeStock"
      parameters:
        - name: "id"
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        '404':
          $ref: "#/components/responses/EntityNotFound"
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerError"
  /v0.1/orders/all:
    get:
      tags:
        - order-controller
      operationId: getOrdersAll
      parameters:
        - name: blockchains
          in: query
          required: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/Blockchain"
        - name: platform
          in: query
          required: false
          schema:
            "$ref": "#/components/schemas/Platform"
        - name: "origin"
          in: query
          required: false
          schema:
            type: string
        - name: "continuation"
          in: query
          required: false
          schema:
            type: string
        - name: "size"
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Orders"
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerError"
  /v0.1/orders/sell/byMaker:
    get:
      tags:
        - order-controller
      operationId: getSellOrdersByMaker
      parameters:
        - name: platform
          in: query
          required: false
          schema:
            "$ref": "#/components/schemas/Platform"
        - name: "maker"
          in: query
          required: true
          schema:
            type: string
        - name: "origin"
          in: query
          required: false
          schema:
            type: string
        - name: "continuation"
          in: query
          required: false
          schema:
            type: string
        - name: "size"
          in: query
          required: false
          schema:
            minimum: 1
            type: integer
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Orders"
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerError"
  /v0.1/orders/sell/byItem:
    get:
      tags:
        - order-controller
      operationId: getSellOrdersByItem
      parameters:
        - name: platform
          in: query
          required: false
          schema:
            "$ref": "#/components/schemas/Platform"
        - name: "contract"
          in: query
          required: true
          schema:
            type: string
        - name: "tokenId"
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/BigInteger"
        - name: "maker"
          in: query
          required: false
          schema:
            type: string
        - name: "origin"
          in: query
          required: false
          schema:
            type: string
        - name: "continuation"
          in: query
          required: false
          schema:
            type: string
        - name: "size"
          in: query
          required: false
          schema:
            minimum: 1
            type: integer
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Orders"
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerError"
  /v0.1/orders/sell/byCollection:
    get:
      tags:
        - order-controller
      operationId: getSellOrdersByCollection
      parameters:
        - name: platform
          in: query
          required: false
          schema:
            "$ref": "#/components/schemas/Platform"
        - name: "collection"
          in: query
          required: true
          schema:
            type: string
        - name: "origin"
          in: query
          required: false
          schema:
            type: string
        - name: "continuation"
          in: query
          required: false
          schema:
            type: string
        - name: "size"
          in: query
          required: false
          schema:
            minimum: 1
            type: integer
            format: int32
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Orders"
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerError"
  /v0.1/orders/sell:
    get:
      tags:
        - order-controller
      operationId: getSellOrders
      parameters:
        - name: "blockchains"
          in: query
          required: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/Blockchain"
        - name: platform
          in: query
          required: false
          schema:
            "$ref": "#/components/schemas/Platform"
        - name: "origin"
          in: query
          required: false
          schema:
            type: string
        - name: "continuation"
          in: query
          required: false
          schema:
            type: string
        - name: "size"
          in: query
          required: false
          schema:
            minimum: 1
            type: integer
            format: int32
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Orders"
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerError"
  /v0.1/orders/bids/byMaker:
    get:
      tags:
        - order-controller
      operationId: getOrderBidsByMaker
      parameters:
        - name: platform
          in: query
          required: false
          schema:
            "$ref": "#/components/schemas/Platform"
        - name: "maker"
          in: query
          required: true
          schema:
            type: string
        - name: "origin"
          in: query
          required: false
          schema:
            type: string
        - name: "continuation"
          in: query
          required: false
          schema:
            type: string
        - name: "size"
          in: query
          required: false
          schema:
            minimum: 1
            type: integer
            format: int32
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Orders"
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerError"
  /v0.1/orders/bids/byItem:
    get:
      tags:
        - order-controller
      operationId: getOrderBidsByItem
      parameters:
        - name: platform
          in: query
          required: false
          schema:
            "$ref": "#/components/schemas/Platform"
        - name: "contract"
          in: query
          required: true
          schema:
            type: string
        - name: "tokenId"
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/BigInteger"
        - name: "maker"
          in: query
          required: false
          schema:
            type: string
        - name: "origin"
          in: query
          required: false
          schema:
            type: string
        - name: "continuation"
          in: query
          required: false
          schema:
            type: string
        - name: "size"
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Orders"
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/ServerError"

  ######################## Activity API ########################
...
components:
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/UnionApiErrorBadRequest"
    ServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/UnionApiErrorServerError"
    EntityNotFound:
      description: Not Found
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/UnionApiErrorEntityNotFound"